requirements.txt
langchain
langchain-openai
langchain-community
langchain-pinecone
python-dotenv
langchainhub